general: moved GameItem from feeder to shared
