+++ ClrVpin/Feeder/FeederResults.xaml
@@ -9,7 +9,8 @@
                     xmlns:shared="clr-namespace:ClrVpin.Models.Shared"
                     xmlns:vps="clr-namespace:ClrVpin.Models.Feeder.Vps"
                     xmlns:feederModels="clr-namespace:ClrVpin.Models.Feeder"
-                    xmlns:feeder="clr-namespace:ClrVpin.Feeder">
+                    xmlns:feeder="clr-namespace:ClrVpin.Feeder"
+                    xmlns:game="clr-namespace:ClrVpin.Models.Shared.Game">
 
   <ResourceDictionary.MergedDictionaries>
     <controls:SharedResourceDictionary Source="\Controls\Styles.xaml" />
@@ -568,10 +569,10 @@
             </b:Interaction.Triggers>
 
             <DataGrid.Columns>
-              <DataGridTextColumn d:DataContext="{d:DesignInstance Type=feederModels:GameItem}" ElementStyle="{StaticResource VerticalAlignDataElementTextStyle}"
+              <DataGridTextColumn d:DataContext="{d:DesignInstance Type=game:GameItem}" ElementStyle="{StaticResource VerticalAlignDataElementTextStyle}"
                                   Header="#" Binding="{Binding Index}" Width="45" />
 
-              <DataGridTemplateColumn d:DataContext="{d:DesignInstance Type=feederModels:GameItem}" CanUserSort="False" CanUserResize="False" CanUserReorder="False" 
+              <DataGridTemplateColumn d:DataContext="{d:DesignInstance Type=game:GameItem}" CanUserSort="False" CanUserResize="False" CanUserReorder="False" 
                                       Width="30" CellStyle="{StaticResource AlignedDataGridCellStyle}">
                 <DataGridTemplateColumn.CellTemplate>
                   <DataTemplate>
@@ -602,15 +603,15 @@
                 </DataGridTemplateColumn.CellTemplate>
               </DataGridTemplateColumn>
 
-              <DataGridTextColumn d:DataContext="{d:DesignInstance Type=feederModels:GameItem}" ElementStyle="{StaticResource VerticalAlignDataElementTextStyle}" Header="Table"
+              <DataGridTextColumn d:DataContext="{d:DesignInstance Type=game:GameItem}" ElementStyle="{StaticResource VerticalAlignDataElementTextStyle}" Header="Table"
                                   Binding="{Binding Name}"
                                   Width="255" />
-              <DataGridTextColumn d:DataContext="{d:DesignInstance Type=feederModels:GameItem}" ElementStyle="{StaticResource VerticalAlignDataElementTextStyle}" Header="Manufacturer"
+              <DataGridTextColumn d:DataContext="{d:DesignInstance Type=game:GameItem}" ElementStyle="{StaticResource VerticalAlignDataElementTextStyle}" Header="Manufacturer"
                                   Binding="{Binding Manufacturer}" Width="160" />
-              <DataGridTextColumn d:DataContext="{d:DesignInstance Type=feederModels:GameItem}" ElementStyle="{StaticResource VerticalAlignDataElementTextStyle}" Header="Year"
+              <DataGridTextColumn d:DataContext="{d:DesignInstance Type=game:GameItem}" ElementStyle="{StaticResource VerticalAlignDataElementTextStyle}" Header="Year"
                                   Binding="{Binding Year}"
                                   Width="55" />
-              <DataGridTextColumn d:DataContext="{d:DesignInstance Type=feederModels:GameItem}" ElementStyle="{StaticResource VerticalAlignDataElementTextStyle}" Header="Updated"
+              <DataGridTextColumn d:DataContext="{d:DesignInstance Type=game:GameItem}" ElementStyle="{StaticResource VerticalAlignDataElementTextStyle}" Header="Updated"
                                   Binding="{Binding UpdatedAt, StringFormat={}{0:dd/MM/yy hh:mm}}" />
             </DataGrid.Columns>
           </DataGrid>
