+++ ClrVpin/Shared/GameCollections.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using ClrVpin.Models.Feeder;
+using ClrVpin.Models.Shared.Game;
 using PropertyChanged;
 using Utils.Extensions;
